requirements.txt
numpy==1.26.4
pytest==8.3.5
python-dotenv==1.0.1
redis==5.2.1
scikit-learn==1.6.1
scikit-surprise==1.1.4
scipy==1.15.2
